requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.3
flask-pydantic
pydantic
SQLAlchemy
requests
spacy-udpipe
hunspell
pyspellchecker
levenshtein
celery
flask-cors
redis
flask-sqlalchemy